Implement abstract classes
Tests

Refs #157

Make suggested changes

Refs #157
