+++ utils/clients/tests/test_icat_client.py
@@ -1,0 +1,39 @@
+"""
+Test ICAT client
+"""
+import unittest
+
+
+class TestICATClient(unittest.TestCase):
+    """
+    Some thought will be required here as we can't use icat credentials
+    """
+    @unittest.skip("Not yet implemented")
+    def test_default_init(self):
+        """ Not yet implemented """
+        self.fail("Not implemented")
+
+    @unittest.skip("Not yet implemented")
+    def test_non_default_init(self):
+        """ Not yet implemented """
+        self.fail("Not implemented")
+
+    @unittest.skip("Not yet implemented")
+    def test_valid_connection(self):
+        """ Not yet implemented """
+        self.fail("Not implemented")
+
+    @unittest.skip("Not yet implemented")
+    def test_invalid_connection(self):
+        """ Not yet implemented """
+        self.fail("Not implemented")
+
+    @unittest.skip("Not yet implemented")
+    def test_stop_connection(self):
+        """ Not yet implemented """
+        self.fail("Not implemented")
+
+    @unittest.skip("Not yet implemented")
+    def test_query_icat(self):
+        """ Not yet implemented """
+        self.fail("Not implemented")
